Правки по карте 1
